requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
yfinance
pydantic-settings
